Revert "ADM-519 refocter: format the code"
This reverts commit 40c1d9f0a22e3ee4297c6af1f396a77d723a186f.
