+++ frontend/__tests__/src/components/Metrics/ConfigStep/SourceControl.test.tsx
@@ -5,7 +5,6 @@ import { SourceControl } from '@src/components/Metrics/ConfigStep/SourceControl'
 import {
   CONFIG_TITLE,
   ERROR_MESSAGE_COLOR,
-  ERROR_PAGE_ROUTE,
   MOCK_SOURCE_CONTROL_URL,
   RESET,
   SOURCE_CONTROL_FIELDS,
@@ -20,6 +19,7 @@ import { setupServer } from 'msw/node'
 import { rest } from 'msw'
 import { HttpStatusCode } from 'axios'
 import { navigateMock } from '../../../../setupTests'
+import { ERROR_PAGE_ROUTE } from '@src/constants'
 import userEvent from '@testing-library/user-event'
 
 export const fillSourceControlFieldsInformation = () => {
