+++ frontend/__tests__/src/components/Metrics/ReportStep/ReportStep.test.tsx
@@ -2,7 +2,6 @@ import { act, render, waitFor } from '@testing-library/react'
 import { ReportStep } from '@src/components/Metrics/ReportStep'
 import {
   BACK,
-  ERROR_PAGE_ROUTE,
   EXPECTED_REPORT_VALUES,
   EXPORT_BOARD_DATA,
   EXPORT_PIPELINE_DATA,
@@ -18,6 +17,7 @@ import { navigateMock } from '../../../../setupTests'
 import mocked = jest.mocked
 import { useExportCsvEffect } from '@src/hooks/useExportCsvEffect'
 import { useGenerateReportEffect } from '@src/hooks/useGenerateReportEffect'
+import { ERROR_PAGE_ROUTE } from '@src/constants'
 
 jest.mock('@src/hooks/useGenerateReportEffect', () => ({
   useGenerateReportEffect: jest.fn().mockReturnValue({
