+++ frontend/__tests__/src/fixtures.ts
@@ -620,10 +620,6 @@ export const CLASSIFICATION_WARNING_MESSAGE = `Some classifications in import da
 export const HOME_VERIFY_IMPORT_WARNING_MESSAGE =
   'The content of the imported JSON file is empty. Please confirm carefully'
 
-export const HOME_PAGE_ROUTE = '/home'
-
-export const ERROR_PAGE_ROUTE = '/error-page'
-
 export const ERROR_PAGE_MESSAGE =
   'Something on internet is not quite right. Perhaps head back to our homepage and try again.'
