+++ frontend/__tests__/src/pages/ErrorPage.test.tsx
@@ -1,25 +1,33 @@
 import { render } from '@testing-library/react'
-import { MemoryRouter } from 'react-router-dom'
+import { BrowserRouter } from 'react-router-dom'
 import ErrorPage from '@src/pages/ErrorPage'
-import { PROJECT_NAME } from '../fixtures'
-import { setupStore } from '../utils/setupStoreUtil'
-import { Provider } from 'react-redux'
+import { ERROR_PAGE_MESSAGE, RETRY_BUTTON } from '../fixtures'
+import userEvent from '@testing-library/user-event'
+import { navigateMock } from '../../setupTests'
+import { HOME_PAGE_ROUTE } from '@src/constants'
 
 describe('error page', () => {
-  const setup = () => {
-    store = setupStore()
-    return render(
-      <Provider store={store}>
-        <MemoryRouter>
-          <ErrorPage />
-        </MemoryRouter>
-      </Provider>
-    )
-  }
-  let store = null
   it('should show error message when render error page', () => {
-    const { getByText } = setup()
+    const { getByText } = render(
+      <BrowserRouter>
+        <ErrorPage />
+      </BrowserRouter>
+    )
 
-    expect(getByText(PROJECT_NAME)).toBeInTheDocument()
+    expect(getByText('OH NO !')).toBeInTheDocument()
+    expect(getByText(ERROR_PAGE_MESSAGE)).toBeInTheDocument()
+    expect(getByText(RETRY_BUTTON)).toBeInTheDocument()
+  })
+
+  it('should go to home page when click button', async () => {
+    const { getByText } = render(
+      <BrowserRouter>
+        <ErrorPage />
+      </BrowserRouter>
+    )
+
+    await userEvent.click(getByText(RETRY_BUTTON))
+
+    expect(navigateMock).toHaveBeenCalledWith(HOME_PAGE_ROUTE)
   })
 })
